feat(bmi-calculator): separate entry point from main.dart
