+++ lib/bmi-calculator/app.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import './main_page.dart';
+
+class App extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'BMI Calculator',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: MainPage(),
+    );
+  }
+}
