+++ lib/bmi-calculator/main_page.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class MainPage extends StatefulWidget {
+  MainPage({Key key}) : super(key: key);
+
+  @override
+  _MainPageState createState() => _MainPageState();
+}
+
+class _MainPageState extends State<MainPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: <Widget>[],
+      ),
+    );
+  }
+}
